% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosimilarity_matrix.R
\name{cosimilarity_matrix}
\alias{cosimilarity_matrix}
\title{Returns cosine cosimilarity matrix for the terms generated by keyclust}
\usage{
cosimilarity_matrix(x)
}
\arguments{
\item{x}{output from \code{\link[=keyclust]{keyclust()}}}
}
\value{
An N x N matrix of cosine cosimilarity values, where n is the number of terms in the provided embedding model
}
\description{
A function that extracts the cosimilarity matrix for
terms generated by \code{\link[=keyclust]{keyclust()}}
}
\keyword{keyclust}
