% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyclust.R
\name{print.keyclust}
\alias{print.keyclust}
\title{Prints terms generated by keyclust}
\usage{
\method{print}{keyclust}(x, ...)
}
\arguments{
\item{x}{output from \code{\link[=keyclust]{keyclust()}}}

\item{...}{additional arguments not used}
}
\value{
A message indicating the number of keywords produced and a preview of the first few keywords.
}
\description{
Prints terms generated by keyclust
}
\keyword{keyclust}
